+++ src/03_commands/01_show-system-info-command.js
@@ -1,5 +1,4 @@
-let calculate_ram_usage = require('./03_calculate_ram_usage.js');
-let get_ram_braille_character = require('./04_get_ram_braille_character.js');
+let { calculate_ram_usage_internal, get_ram_braille_character } = require('../01_monitors/02_ram-monitor.js');
 let vscode = require('vscode');
 let os = require('os');
 
@@ -14,7 +13,7 @@ async function show_system_info_command() {
     let cpu_info = os.cpus();
     let cpu_model = cpu_info[0].model;
     let cpu_core_count = cpu_info.length;
-    let ram_usage_info = await calculate_ram_usage();
+    let ram_usage_info = await calculate_ram_usage_internal();
     let operating_system_platform = os.platform();
 
     //